+++ R/plot.R
@@ -6,10 +6,7 @@ plot.errors <- function(x, y, ...) {
     x <- seq_along(x)
   }
 
-  xlim <- c(errors_min(min(x)), errors_max(max(x)))
-  ylim <- c(errors_min(min(y)), errors_max(max(y)))
-
-  NextMethod(xlim=xlim, ylim=ylim)
+  NextMethod(xlim=range(x), ylim=range(y))
 
   if (inherits(x, "errors"))
     graphics::segments(errors_min(x), y, errors_max(x), y)