+++ R/errors.R
@@ -1,31 +1,43 @@
 #' \pkg{errors}: Error Propagation for R Vectors
 #'
 #' Support for painless automatic error propagation in numerical operations.
+#' Errors are automatically propagated when you operate with \code{errors}
+#' objects, or with \code{errors} and numeric objects (then numeric values are
+#' automatically coerced to errors with zero error).
+#' \cr\cr This package treats errors as coming from Gaussian, linear and independent
+#' sources, and propagates them using the first-order Taylor series method for
+#' propagation of uncertainty. Although the above assumptions are valid in a wide
+#' range of applications in science and engineering, the practitioner should
+#' evaluate whether they apply for each particular case.
 #'
 #' @author Iñaki Ucar
 #'
-#' @return \code{errors} returns a vector of errors. \code{errors_max}
-#' (\code{errors_min}) returns the object plus (minus) the errors.
+#' @details \code{errors} returns a vector of errors. \code{errors_max}
+#' (\code{errors_min}) returns the values plus (minus) the errors.
 #'
-#' \code{`errors<-`} sets the error values and converts \code{x} into an object
-#' of class \code{errors}. \code{set_errors} is a pipe-friendly version of
+#' \code{`errors<-`} sets the error values (and converts \code{x} into an object
+#' of class \code{errors}). \code{set_errors} is a pipe-friendly version of
 #' \code{`errors<-`} and returns an object of class \code{errors}. \code{as.errors}
-#' is an alias for \code{set_errors} with \code{value = 0} by default.
+#' is an alias for \code{set_errors}.
 #'
-#' @details Errors are automagically propagated when you operate with \code{errors}
-#' objects, or with \code{errors} and numeric objects. By default, propagation
-#' is made using a first-order Taylor series expansion.
+#' @seealso
+#' \code{\link{groupGeneric.errors}}, \code{\link{mean.errors}}.
+#' \code{\link{Extract.errors}}, \code{\link{c}}, \code{\link{rep}}, \code{\link{cbind.errors}}.
+#' \code{\link{format.errors}}, \code{\link{print.errors}}, \code{\link{plot.errors}}.
+#' \code{\link{as.data.frame.errors}}, \code{\link{as.matrix.errors}}, \code{\link{t}}.
 #'
 #' @examples
 #' x = 1:3
 #' class(x)
+#' x
 #' errors(x) <- 0.1
 #' class(x)
 #' x
-#' errors(x) <- seq(0.1, 0.3, 0.1)
-#' x
-#' # numeric values are automatically coerced to errors
-#' x^2 + x
+#'
+#' (x <- set_errors(x, seq(0.1, 0.3, 0.1)))
+#' errors_max(x)
+#' errors_min(x)
+#'
 #' @docType package
 #' @import stats
 #' @name errors