+++ R/plot.R
@@ -1,5 +1,17 @@
-# nocov start
+#' Generic X-Y Plotting
+#'
+#' S3 method for \code{errors} objects which automatically prints the error bars.
+#'
+#' @inheritParams graphics::plot
+#' @param ... additional arguments (see \code{\link[graphics]{plot}}).
+#'
+#' @examples
+#' cars <- as.matrix(cars)
+#' cars <- as.data.frame(set_errors(cars, cars * 0.05))
+#' plot(cars)
+#'
 #' @export
+# nocov start
 plot.errors <- function(x, y, ...) {
   if (missing(y)) {
     y <- x